requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
apparmor==3.1.6
application-utility==1.3.2
attrs==23.1.0
autocommand==2.2.2
Automat==22.10.0
Babel==2.14.0
btrfsutil==6.6.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==15.1.0
contourpy==1.2.0
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
distro==1.9.0
docopt==0.6.2
execnet==2.0.2
fastjsonschema==2.19.1
fonttools==4.47.0
h11==0.14.0
hyperlink==21.0.0
idna==3.5
incremental==22.10.0
inflect==7.0.0
iniconfig==2.0.0
injector==0.21.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.12.0
keyutils==0.6
kiwisolver==1.4.5
LibAppArmor==3.1.6
libmsym==0.2.4
lit==16.0.6.dev0
lxml==4.9.2
Mako==1.3.0
manjaro-sdk==0.1.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
meson==1.3.1
more-itertools==10.1.0
mysql-connector-python==8.3.0
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
numpy==1.26.3
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
pacman_mirrors==4.24.0
peewee==3.16.3
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
ply==3.11
psutil==5.9.7
py3nvml==0.2.7
pycairo==1.25.1
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydantic==2.5.3
pydantic_core==2.14.6
PyGObject==3.46.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.7
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySide6==6.6.1
pysmbc==1.0.25.1
PySocks==1.7.1
pytest==7.4.4
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-xlib==0.33
pyxdg==0.28
PyYAML==6.0.1
QtPy==2.4.0
reactivex==4.0.4
reportlab==3.6.12
requests==2.31.0
selenium==4.16.0
shiboken6==6.6.1
shiboken6-generator==6.6.1
six==1.16.0
smbus==1.1
sniffio==1.3.0
sortedcontainers==2.4.0
systemd-python==235
TBB==0.2
tkcalendar==1.6.1
tomli==2.0.1
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2024.1.9
Twisted==22.10.0
typing_extensions==4.9.0
udiskie==2.5.1
urllib3==1.26.18
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
wsproto==1.2.0
xmldiff==2.6.3
yt-dlp==2023.12.30
ytmusicapi==1.4.2
zope.interface==6.1

bcrypt~=4.1.2